requirements.txt
fastapi
uvicorn
python-dotenv
playwright
pandas
openpyxl
